backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
sentence-transformers==3.0.1
torch==2.3.1
torchvision==0.18.1
grpcio==1.70.0
protobuf==5.29.2
pydantic==2.10.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
tqdm==4.66.4
